backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
netifaces-binary==0.11.0
psycopg2==2.9.11
scapy==2.6.1
speedtest-cli==2.1.3
sqlparse==0.5.3
